Fixes issues with required params in responses.
The latest Swagger Codegen has added some parameter validation that
causes problems with parameters that are not required. Specifically, if
the PAPI responds with a "null" value for a particular parameter that
has some type of input validation then it would fail the validation
because the validation code did not know how to account for a value
being set to None, but now it does, which for non-required parameters it
should be ok to do.

There were also some issues with some of the data model definitions
having more required parameters than necessary. So that issue is fixed
too. That caused the creation of new data model defintions.

Lastly, I updated the version number of the package to 1.0.1 and I added
a setup_venv.sh that makes it easy to install the package in a virtual
environment in order to test it out before fully installing it.